backend/requirements.txt
fastapi
motor
openai
pinecone
python-dotenv
uvicorn
